Enable ESLint in CI and refactor core code
Add an ESLint config and run lint as part of the CI workflow. Update package/build config to support linting and make related changes across the codebase (core embedding/search engines, SQLiteStore modules, workers, UI, and tests). Remove the MCPServer.ts file and adjust tests/integration expectations to match refactors. These changes introduce linting enforcement and code cleanups/refactors to improve maintainability and search behavior.

+++ src/sqlJsBundled.ts
@@ -1,3 +1,3 @@
 import sqlJsWasm from "sql.js/dist/sql-wasm.wasm";
 
-export const bundledSqlJsWasm = sqlJsWasm as Uint8Array;
+export const bundledSqlJsWasm = sqlJsWasm;
